+++ Assets/Prefabs/Pacmom/Dust.prefab
@@ -119,6 +119,7 @@ GameObject:
   - component: {fileID: 8324175502395817452}
   - component: {fileID: 5395930438513632476}
   - component: {fileID: 5522138042445295253}
+  - component: {fileID: 8517634723861635253}
   m_Layer: 7
   m_Name: Dust
   m_TagString: Untagged
@@ -218,6 +219,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   gameController: {fileID: 0}
+  <dustID>k__BackingField: 0
 --- !u!114 &8324175502395817452
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -265,6 +267,21 @@ MonoBehaviour:
   movement: {fileID: 8324175502395817452}
   inside: {fileID: 0}
   outside: {fileID: 0}
+  crossRoad: {fileID: 0}
+--- !u!114 &8517634723861635253
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5555307432301182834}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 3656dd137e2f5154599b72ced36ff159, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  dialogue: {fileID: 0}
+  reachTime: 0
 --- !u!1 &8942196247262051131
 GameObject:
   m_ObjectHideFlags: 0
