+++ Assets/Scripts/Runtime/CH1/Pacmom/Dust.cs
@@ -5,18 +5,11 @@ namespace Runtime.CH1.Pacmom
 {
     public class Dust : MonoBehaviour
     {
-        [SerializeField]
-        private InGameDialogue dialogue;
-
         public PMGameController gameController;
         public MovementAndEyes movement { get; private set; }
         public AI ai { get; private set; }
-
         [field:SerializeField]
         public int dustID { get; private set; }
-        [SerializeField]
-        private float reachTime = 0f;
-        private bool dustTalked = false;
 
         private void Awake()
         {
@@ -74,27 +67,6 @@ namespace Runtime.CH1.Pacmom
                 else
                     gameController?.DustEaten(this);
             }
-        }
-
-        private void OnCollisionStay2D(Collision2D collision)
-        {
-            if (collision.gameObject.layer == LayerMask.NameToLayer(GlobalConst.PlayerStr))
-            {
-                reachTime += Time.deltaTime;
-            }
-
-            if (!dustTalked && reachTime > 1.2f)
-            {
-                if (ai.isStronger)
-                    dialogue.BlockedDialogue(dustID);
-                dustTalked = true;
-            }
-        }
-
-        private void OnCollisionExit2D(Collision2D collision)
-        {
-            reachTime = 0f;
-            dustTalked = false;
         }
     }
 }
