+++ Assets/Scripts/Runtime/CH1/Pacmom/DustBlocked.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3656dd137e2f5154599b72ced36ff159
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
